resolves advanced search button
