+++ awx/ui/src/components/HealthCheckButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './HealthCheckButton';
