+++ awx/ui/src/screens/Instances/InstanceList/InstanceList.js
@@ -1,8 +1,8 @@
 import React, { useCallback, useEffect } from 'react';
-
-import { Plural, t } from '@lingui/macro';
+import { t } from '@lingui/macro';
 import { useLocation } from 'react-router-dom';
 import 'styled-components/macro';
+import { PageSection, Card } from '@patternfly/react-core';
 
 import useExpanded from 'hooks/useExpanded';
 import DataListToolbar from 'components/DataListToolbar';
@@ -13,13 +13,11 @@ import PaginatedTable, {
 } from 'components/PaginatedTable';
 import AlertModal from 'components/AlertModal';
 import ErrorDetail from 'components/ErrorDetail';
-
 import useRequest, { useDismissableError } from 'hooks/useRequest';
 import useSelected from 'hooks/useSelected';
 import { InstancesAPI } from 'api';
 import { getQSConfig, parseQueryString } from 'util/qs';
-
-import { Button, Tooltip, PageSection, Card } from '@patternfly/react-core';
+import HealthCheckButton from 'components/HealthCheckButton';
 import InstanceListItem from './InstanceListItem';
 
 const QS_CONFIG = getQSConfig('instance', {
@@ -86,31 +84,6 @@ function InstanceList() {
   const { expanded, isAllExpanded, handleExpand, expandAll } =
     useExpanded(instances);
 
-  const hopNodeSelected = selected.filter(
-    (instance) => instance.node_type === 'hop'
-  ).length;
-
-  const buildTooltip = () => {
-    if (hopNodeSelected) {
-      return (
-        <Plural
-          value={hopNodeSelected}
-          one="Cannot run health check on a hop node.  Deselect the hop node to run a health check."
-          other="Cannot run health check on hop nodes.  Deselect the hop nodes to run health checks."
-        />
-      );
-    }
-    return selected.length ? (
-      <Plural
-        value={selected.length}
-        one="Click to run a health check on the selected instance."
-        other="Click to run a health check on the selected instances."
-      />
-    ) : (
-      t`Select an instance to run a health check.`
-    );
-  };
-
   return (
     <>
       <PageSection>
@@ -157,18 +130,11 @@ function InstanceList() {
                 onExpandAll={expandAll}
                 qsConfig={QS_CONFIG}
                 additionalControls={[
-                  <Tooltip ouiaId="healthCheckTooltip" content={buildTooltip()}>
-                    <div>
-                      <Button
-                        isDisabled={
-                          !selected.length || Boolean(hopNodeSelected)
-                        }
-                        variant="secondary"
-                        ouiaId="health-check"
-                        onClick={fetchHealthCheck}
-                      >{t`Health Check`}</Button>
-                    </div>
-                  </Tooltip>,
+                  <HealthCheckButton
+                    onClick={fetchHealthCheck}
+                    isDisabled={!selected.length || Boolean(hopNodeSelected)}
+                    selectedItems={selected}
+                  />,
                 ]}
               />
             )}
